Began implementing GameObject and Item
Instead of having trees and ores as tiles, they will be game objects
that can be placed on tiles. This will allow them to be placed on a
variety of terrain instead of just grass. This will also allow game
objects to be harvested if they have a resource. One downside of this
method is that a new game object map file will be needed and memory
usage will be increased.